Update media query, file names, copywrite
Media query is now reorganized to match the sequential appearance of
classes and elements in the html. File names for codec portraits were
updated to match the naming scheme of the rest of the html document.
Copywrite was updated to reflect ownership of the Metal Gear IP by
Konami and development of the website by me